line_bot/requirements.txt
openai==0.28.0
langchain==0.0.292
line-bot-sdk==3.5.0
